add AuthGuard for items handlers

+++ src/items/items.service.ts
@@ -5,6 +5,7 @@ import { InjectRepository } from '@nestjs/typeorm'
 import { Item } from './item.entity'
 import { ItemStatus } from './item-status.enum'
 import { GetItemsFilterDto } from './dto/get-items-filter.dto'
+import { User } from '../auth/user.entity'
 
 @Injectable()
 export class ItemsService {
@@ -13,12 +14,12 @@ export class ItemsService {
     private itemRepository: ItemRepository,
   ) {}
 
-  async getAllItems(filterDto: GetItemsFilterDto): Promise<Item[]> {
-    return this.itemRepository.getAllItems(filterDto)
+  async getAllItems(filterDto: GetItemsFilterDto, user: User): Promise<Item[]> {
+    return this.itemRepository.getAllItems(filterDto, user)
   }
 
-  async getItemById(id: number): Promise<Item> {
-    const found = await this.itemRepository.findOne(id)
+  async getItemById(id: number, user: User): Promise<Item> {
+    const found = await this.itemRepository.findOne({ where: { id, userId: user.id } })
     if (!found) {
       throw new NotFoundException(`Item with id ${id} not found`)
     }
@@ -26,20 +27,21 @@ export class ItemsService {
     return found
   }
 
-  async createItem(createItemDto: CreateItemDto): Promise<Item> {
-    return this.itemRepository.createItem(createItemDto)
+  async createItem(createItemDto: CreateItemDto, user: User): Promise<Item> {
+    return this.itemRepository.createItem(createItemDto, user)
   }
 
-  async deleteItem(id: number): Promise<void> {
-    const result = await this.itemRepository.delete(id)
+  async deleteItem(id: number, user: User): Promise<void> {
+    const result = await this.itemRepository.delete({ id, userId: user.id })
 
     if (result.affected === 0) {
       throw new NotFoundException(`Item with id ${id} not found`)
     }
   }
 
-  async updateItemStatus(id: number, status: ItemStatus): Promise<Item> {
-    const item = await this.getItemById(id)
+  async updateItemStatus(id: number, status: ItemStatus, user: User): Promise<Item> {
+    const item = await this.getItemById(id, user)
+
     item.status = status
     await item.save()
